refactor to keep it clear
